writing

+++ idl/message.thrift
@@ -8,7 +8,7 @@ enum ErrCode {
 }
 
 struct BaseResp {
-    1: i64 status_code
+    1: i32 status_code
     2: string status_message
     3: i64 service_time
 }
